+++ libs/dot-components/src/lib/components/auto-complete/AutoComplete.stories.tsx
@@ -1,23 +1,38 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { Story, Meta } from '@storybook/react/types-6-0';
 
-import { DotAutoComplete, AutoCompleteProps } from './AutoComplete';
+import {
+  DotAutoComplete,
+  AutoCompleteProps,
+  parseAutoCompleteValue,
+} from './AutoComplete';
+
+const batman = { group: 'D.C.', title: 'Batman', error: true };
+
+const defaultValueWithError = [batman];
 
 export default {
-  title: 'Experimental/Auto Complete',
+  title: 'Components/Auto Complete',
   component: DotAutoComplete,
   argTypes: {
+    inputId: {
+      defaultValue: 'input-id',
+    },
     options: {
       defaultValue: [
-        { category: 'Marvel', title: 'Hulk' },
-        { category: 'Marvel', title: 'Thor' },
-        { category: 'Marvel', title: 'Ironman' },
-        { category: 'Marvel', title: 'Spiderman' },
-        { category: 'D.C.', title: 'Batman' },
-        { category: 'D.C.', title: 'Flash' },
-        { category: 'D.C.', title: 'Aquaman' },
-        { category: 'D.C.', title: 'Wonderwoman' },
+        { group: 'D.C.', title: 'Aquaman', error: false },
+        batman,
+        { group: 'D.C.', title: 'Flash', error: false },
+        { group: 'Marvel', title: 'Hulk', error: false },
+        { group: 'Marvel', title: 'Ironman', error: false },
+        { group: 'D.C.', title: 'Robin', error: false },
+        { group: 'Marvel', title: 'Spiderman', error: false },
+        { group: 'Marvel', title: 'Thor', error: false },
+        { group: 'D.C.', title: 'Wonderwoman', error: false },
       ],
+    },
+    placeholder: {
+      defaultValue: 'Select a hero',
     },
   },
 } as Meta;
@@ -25,3 +40,25 @@ export default {
 export const Default: Story<AutoCompleteProps> = (args) => (
   <DotAutoComplete {...args} />
 );
+
+export const WithError: Story<AutoCompleteProps> = (args) => {
+  const [helperText, setHelperText] = useState('No Batman without Robin!');
+  const [error, setError] = useState(true);
+  const onChange = (_event, value, reason) => {
+    const parsedValue = parseAutoCompleteValue(value);
+    batman.error =
+      parsedValue.indexOf('Batman') !== -1 &&
+      parsedValue.indexOf('Robin') === -1;
+    setError(batman.error);
+    setHelperText(batman.error ? 'No Batman without Robin!' : null);
+  };
+  return (
+    <DotAutoComplete
+      defaultValue={defaultValueWithError}
+      error={error}
+      helperText={helperText}
+      onChange={onChange}
+      {...args}
+    />
+  );
+};